DDA-55: Rename open room to add connection button
