+++ src/react/features/connectors/folder/buttons/buttons.js
@@ -1,9 +1,9 @@
-export const OPEN_ROOM_BUTTON = 'OPEN_ROOM_BUTTON'
+export const ADD_CONNECTION_BUTTON = 'ADD_CONNECTION_BUTTON'
 
 export const buttons = [
   {
-    id: OPEN_ROOM_BUTTON,
-    text: 'Open room',
+    id: ADD_CONNECTION_BUTTON,
+    text: 'Add connection',
     data: {
       icon: 'folder-open',
     },
